Enhance API documentation with Swagger annotations for Admin, Reservation, and Login controllers. Update JWT expiration time to a longer duration and add Swagger token generation component.

+++ src/main/java/com/aly/brightskies/task3/security/JWTUtility.java
@@ -19,7 +19,7 @@ public class JWTUtility {
     @Value("${jwt.secret}")
     private String secret;
     @Value("${jwt.expiration}")
-    private int expirationTime;
+    private Long expirationTime;
     private Key key;
     @Autowired
     private UserRepo userRepo;
